+++ services/ItemCountable.java
@@ -2,12 +2,12 @@ package services;
 
 
 public class ItemCountable extends Item{
-    public ItemCountable(String name, Double price, Double amount){
+    public ItemCountable(String name, int price, int amount){
         this.name = name;
         this.price = price;
         this.amount = amount;
     }
-    public void addAmount(double amountToAdd){
+    public void addAmount(int amountToAdd){
         this.amount+=amountToAdd;
     }
