+++ utility/ItemDeserializer.java
@@ -1,6 +1,5 @@
 package utility;
 
-import com.fasterxml.jackson.core.JacksonException;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
@@ -11,7 +10,6 @@ import services.ItemUncountable;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.List;
 
 public class ItemDeserializer  extends JsonDeserializer<Item> {
 
@@ -23,7 +21,7 @@ public class ItemDeserializer  extends JsonDeserializer<Item> {
 
 
             String name = node.get("name").asText();
-            double price = node.get("price").asDouble();
+            int price = node.get("price").asInt();
             ArrayList<Double> weight = new ArrayList<>();
             JsonNode weightList = node.get("weightList");
             for(JsonNode value : weightList){
@@ -33,8 +31,8 @@ public class ItemDeserializer  extends JsonDeserializer<Item> {
         }
         else if(node.has("amount")){
             String name = node.get("name").asText();
-            double price = node.get("price").asDouble();
-            double amount = node.get("amount").asDouble();
+            int price = node.get("price").asInt();
+            int amount = node.get("amount").asInt();
             return new ItemCountable(name,price,amount);
         }
